docs(readme): add three badges to top of readme (license, contributors, pull requests)
Added three badges to the top of the README with useful information.

- License: display the type of license used on this project with a link
  to the license text.
- Contributors: display the total number of contributors on this
  repository
- Pull requests: display the total number of pull requests that have
  been closed on this repository

The goal of the license badge is to avoid the need to add an entire
"license" section to the README. The goal of the other two badges is to
hopefully increase engagement by encouraging community members to
contribute to our project so they can watch the number of contributors
and PRs increase in real time.

re #64